added and tested user creation, growl notification, ssh and swupd

+++ manifests/classes/swupd.pp
@@ -2,6 +2,7 @@
 
 class swupd {
 	exec { swupd:
+	onlyif => "/usr/sbin/softwareupdate -l | grep 'No updates'",
 	command => "/usr/sbin/softwareupdate -i -a && reboot",
 	logoutput => true,
 	}
